requirements.txt
ultralytics>=8.2.0
torch>=2.1.0
torchvision>=0.16.0
opencv-python
albumentations>=1.3.1
pyyaml
tqdm
numpy
pandas
matplotlib
scikit-learn
wandb>=0.16.0
optuna>=3.5.0
pycocotools; platform_system != "Windows"
loguru
rich
termcolor
